+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 
 ### Breaking changes
 
+* nostr: update `Tags::new` signature ([Yuki Kishimoto])
+
 ### Changed
 
 * nostr: manually impl eq and cmp traits for `RelayUrl` ([Yuki Kishimoto])
@@ -51,6 +53,7 @@
 
 * nostr: add `EventBuilder::allow_self_tagging` ([Yuki Kishimoto])
 * nostr: add `push`, `pop`, `insert`, `remove` and `extend` methods to `Tags` struct ([Yuki Kishimoto])
+* nostr: add `Tags::from_list` ([Yuki Kishimoto])
 * pool: event verification cache ([Yuki Kishimoto])
 * ffi: add Mac Catalyst support in Swift package ([Yuki Kishimoto])